LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/fetch/api/basic/keepalive-expected.txt


PASS same-origin; setting up 
PASS cross-origin redirect; setting up 
PASS cross-origin redirect with preflight; setting up 
FAIL same-origin assert_unreached: Error: assert_equals: expected (string) "on" but got (object) null Reached unreachable code
FAIL cross-origin redirect assert_unreached: Error: assert_equals: expected (string) "on" but got (object) null Reached unreachable code
FAIL cross-origin redirect with preflight assert_unreached: Error: assert_equals: expected (string) "on" but got (object) null Reached unreachable code
